Move DeepWater MOJO into an extension
