+++ build.gradle
@@ -95,7 +95,8 @@ ext {
       project(':h2o-security'),
       project(':h2o-logger'),
       project(':h2o-genmodel-extensions:jgrapht'),
-      project(':h2o-k8s')
+      project(':h2o-k8s'),
+      project(':h2o-genmodel-extensions:deepwater')
     ]
 
     javaProjects = [
@@ -130,7 +131,8 @@ ext {
       project(':h2o-security'),
       project(':h2o-logger'),
       project(':h2o-genmodel-extensions:jgrapht'),
-      project(':h2o-k8s')
+      project(':h2o-k8s'), 
+      project(':h2o-genmodel-extensions:deepwater')
     ]
 
     scalaProjects = [
