+++ h2o-genmodel-extensions/deepwater/build.gradle
@@ -1,0 +1,13 @@
+description = "H2O GenModel Deepwater Extension"
+
+sourceCompatibility = 1.7
+targetCompatibility = 1.7
+
+dependencies {
+    compile project(":h2o-genmodel")
+  // DeepWater dependencies - users will need to provide excluded & compileOnly dependencies
+  compile("ai.h2o:deepwater-backend-api:1.0.4") {
+    exclude group: 'org.slf4j'
+  }
+  compileOnly "com.google.protobuf.nano:protobuf-javanano:3.1.0"
+}
