+++ h2o-genmodel-extensions/deepwater/src/main/java/hex/genmodel/algos/deepwater/DWTextConverter.java
@@ -1,6 +1,9 @@
-package hex.genmodel.easy;
+package hex.genmodel.algos.deepwater;
 
 import hex.genmodel.GenModel;
+import hex.genmodel.easy.CategoricalEncoder;
+import hex.genmodel.easy.EasyPredictModelWrapper;
+import hex.genmodel.easy.RowToRawDataConverter;
 import hex.genmodel.easy.exception.PredictException;
 
 import java.util.Map;
