+++ h2o-genmodel-extensions/deepwater/src/main/resources/META-INF/services/hex.genmodel.ModelMojoReader
@@ -1,0 +1,1 @@
+hex.genmodel.algos.deepwater.DeepwaterMojoReader
