+++ h2o-genmodel/build.gradle
@@ -33,9 +33,4 @@ dependencies {
   compile "net.sf.opencsv:opencsv:2.3"
   compile "com.google.code.gson:gson:2.6.2"
   compile "ai.h2o:h2o-tree-api:0.3.9"
-  // DeepWater dependencies - users will need to provide excluded & compileOnly dependencies
-  compile("ai.h2o:deepwater-backend-api:1.0.4") {
-    exclude group: 'org.slf4j'
-  }
-  compileOnly "com.google.protobuf.nano:protobuf-javanano:3.1.0"
 }
