+++ settings.gradle
@@ -39,6 +39,7 @@ include 'h2o-security'
 include 'h2o-logger'
 include 'h2o-genmodel-extensions:jgrapht'
 include 'h2o-k8s'
+include 'h2o-genmodel-extensions:deepwater'
 
 // GRPC support
 if ("true".equals(System.getenv("H2O_BUILD_GRPC"))) {
